Sutvarkytos smulkios klaidos kode

+++ headers/klase.h
@@ -5,7 +5,7 @@ private:
     std::string vardas;
     std::string pavarde;
     double egzaminas;
-    std::vector<double> nd;
+    std::vector<int> nd;
     double galutinisSuVidurkiu;
     double galutinisSuMediana;
 
@@ -16,16 +16,16 @@ public:
     void setVardas(const string &var) { vardas = var ;};
     void setPavarde(const string &pav) { pavarde = pav; };
     void setEgzaminas(const double &egz) { egzaminas = egz ;};
-    void setND(const vector<double> &nd_) { nd = nd_; };
-    void setPazymys(const double &paz) { nd.push_back(paz) ;};
+    void setND(const vector<int> &nd_) { nd = nd_; };
+    void setPazymys(const int &paz) { nd.push_back(paz) ;};
     void setGalutinisSuVidurkiu(const double &vid) { galutinisSuVidurkiu = vid; };
     void setGalutinisSuMediana(const double &med) { galutinisSuMediana = med ;};
     // getteriai
     string getVardas() const { return vardas; };
     string getPavarde() const { return pavarde; };
     double getEgzaminas() const { return egzaminas; };
-    vector<double> getND() const { return nd ;};
-    double getPazymys(int &i) const { return nd.at(i); };
+    vector<int> getND() const { return nd ;};
+    int getPazymys(int &i) const { return nd.at(i); };
     double getGalutinisSuVidurkiu() const { return galutinisSuVidurkiu ;};
     double getGalutinisSuMediana() const { return galutinisSuMediana ;};
 };
